+++ SpatialGDK/Source/SpatialGDK/Private/SpatialGDKSettings.cpp
@@ -82,9 +82,9 @@ void CheckCmdLineOverrideOptionalString(const TCHAR* CommandLine, const TCHAR* P
 void CheckCmdLineOverrideOptionalStringWithCallback(const TCHAR* CommandLine, const TCHAR* Parameter, const TCHAR* PrettyName,
 													TFunctionRef<void(const FString& Setting)> Callback)
 {
+	TOptional<FString> OverrideValue;
 #if ALLOW_SPATIAL_CMDLINE_PARSING
 	FString TempStr;
-	TOptional<FString> OverrideValue;
 	if (FParse::Value(CommandLine, Parameter, TempStr) && TempStr[0] == '=')
 	{
 		OverrideValue = TempStr.Right(TempStr.Len() - 1); // + 1 to skip =
@@ -150,6 +150,7 @@ USpatialGDKSettings::USpatialGDKSettings(const FObjectInitializer& ObjectInitial
 	, EventTracingSamplingSettingsClass(UEventTracingSamplingSettings::StaticClass())
 	, MaxEventTracingFileSizeBytes(DefaultEventTracingFileSize)
 	, bEnableAlwaysWriteRPCs(false)
+	, bEnableInitialOnlyReplicationCondition(false)
 {
 	DefaultReceptionistHost = SpatialConstants::LOCAL_HOST;
 	RPCRingBufferSizeOverrides.Add(ERPCType::ServerAlwaysWrite, 1);