+++ SpatialGDK/Source/SpatialGDKFunctionalTests/Private/SpatialTest1x2GridSmallInterestStrategy.cpp
@@ -1,8 +1,0 @@
-// Copyright (c) Improbable Worlds Ltd, All Rights Reserved
-
-#include "SpatialTest1x2GridSmallInterestStrategy.h"
-
-USpatialTest1x2GridSmallInterestStrategy::USpatialTest1x2GridSmallInterestStrategy()
-{
-	InterestBorder = 150.0f;
-}